requirements.txt
mcp-server>=0.4.0
httpx>=0.24.0
python-dotenv>=1.0.0